+++ frontend/src/app/components/login/login.components.scss
@@ -1,0 +1,127 @@
+::ng-deep {
+
+  #logInButton .p-button,
+  #registerButton .p-button {
+    background-color: var(--primary-color);
+    color: var(--primary-text-color);
+    border-color: var(--primary-color);
+  }
+
+  #logInButton .p-button:hover,
+  #registerButton .p-button:hover {
+    background-color: var(--accent-color);
+    border-color: var(--accent-color) !important;
+    color: var(--accent-text-color);
+  }
+}
+
+/* 1. Mobile (sm) styles */
+@media (min-width: 320px) and (max-width: 640px) {
+  ::ng-deep {
+    .p-password {
+      height: 44.1px !important;
+    }
+
+    p-button {
+      .p-button-icon-only {
+        width: 25px;
+        height: 25px;
+
+        span {
+          font-size: 18px;
+        }
+      }
+
+      button {
+        width: 100%;
+        height: 100%;
+
+        span {
+          font-size: 10px;
+        }
+      }
+    }
+  }
+
+  form {
+    width: 280px;
+
+    div {
+      font-size: 25px;
+
+      div {
+        width: 200px;
+
+        p,
+        a {
+          font-size: 10px;
+        }
+
+        a {
+          width: 80px;
+          height: 40px;
+        }
+      }
+
+      .p-floatlabel,
+      small {
+        font-size: 10px;
+        width: 200px;
+
+        input {
+          font-size: 15px;
+        }
+
+        label {
+          font-size: 10px;
+        }
+      }
+    }
+  }
+}
+
+/* 2. Mobile (sm) styles */
+@media (max-width: 640px) {
+  .p-floatlabel {
+    font-size: 0.875rem;
+  }
+
+  .p-button {
+    width: 100%;
+  }
+}
+
+/* 3. Tablet (md) styles */
+@media (min-width: 640px) and (max-width: 1024px) {
+  .p-floatlabel {
+    font-size: 1rem;
+  }
+
+  .p-button {
+    width: auto;
+  }
+}
+
+/* 4. Desktop (lg) styles */
+@media (min-width: 1024px) {
+  .p-floatlabel {
+    font-size: 1.125rem;
+  }
+
+  .p-button {
+    width: auto;
+  }
+
+  ::ng-deep {
+    p-button {
+      .p-button-icon-only {
+        width: 45px;
+        height: 45px;
+
+        span {
+          font-size: 35px;
+        }
+      }
+    }
+  }
+}
